norminette + func splited done
