+++ Makefile
@@ -7,6 +7,8 @@ CFLAGS = -Wall -Wextra -Werror -g3 -Iincludes -Isrc -MMD
 SRCS = src/main.c \
 	src/exec/child.c src/exec/heredoc.c src/exec/signal.c \
 	src/exec/env_create.c src/exec/env_manage.c \
+	src/exec/exec.c src/exec/exec_utils.c src/exec/init_prompt.c \
+	src/exec/minishell_utils.c src/exec/minishell_loop.c \
 	src/parsing/get_msh.c src/parsing/parsing_utils.c \
 	src/parsing/fill_token.c src/parsing/token_utils.c \
 	src/parsing/fill_msh.c src/parsing/msh_utils.c \
