+++ includes/libft/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: wnocchi <wnocchi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/08 15:35:35 by wnocchi           #+#    #+#             */
-/*   Updated: 2023/11/11 23:20:38 by wnocchi          ###   ########.fr       */
+/*   Updated: 2024/08/10 13:11:06 by wnocchi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
